Create README.md
DR_DME
We reconstructed a retinal disease dataset based on the Messidor database, consists of 1200 eye fundus color numerical images of the posterior pole acquired from three ophthalmologic departments. In the reconstructed dataset, six classes were labeled, represented by classes 0, 1, 2, 3, 4, and 5, respectively. Class 0 represents the normal images without any lesions; class 1 stands for the images labeled as DR grade 1 and DME grade 0; class 2 indicates the images labeled as DME grade 0 and DR grade 2. Since there are a few images labeled in both DME grade 1 or 2 and DR grade 1 or 2, the images both with DR grade 1 or 2 and DME grade 1 or 2 are regarded as one grade, represented by class 3; class 4 stands for the DR grade 3 and DME grade 0; class 5 signifies the DR grade 3 and DME grade 1 and 2. Table 1 shows the numbers of images in the six classes for our constructed data set. Totally, there are 546 images of class 0, 142 images of class 1, 182 images of class 2, 74 images of class 3, 104 images of class 4 and 152 images of class 5.
Users of the messidor database are encouraged to reference the following article:
                                              Decencière et al.. Feedback on a publicly distributed database: the Messidor database.
                                              Image Analysis & Stereology, v. 33, n. 3, p. 231-234, aug. 2014. ISSN 1854-5165.
                                             Available at: http://www.ias-iss.org/ojs/IAS/article/view/1155 or http://dx.doi.org/10.5566/ias.1155.

Download
You can download our constructed dataset and the code of models from Baidu Drive.
* data on Baidu Drive: 链接：https://pan.baidu.com/s/1ic2SJJa8vaV_hLbkCLcj6A 提取码：43rq
* models on Baidu drive: 链接：https://pan.baidu.com/s/1eFsbWDy193Ua2rpgH_Immg 提取码：72cp